pyproject.toml
[project]
name = "mcp-server-tavily"
version = "0.1.0"
description = "MCP Server providing a tool for Tavily web search."
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "mcp>=1.6.0,<2.0.0",
    "starlette>=0.37.0",
    "uvicorn[standard]>=0.29.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0",
    "langchain-tavily>=0.1.5",
    "langchain-mcp-adapters>=0.0.9",
    "langgraph>=0.0.20",
    "langchain-together>=0.0.5",
]


[dependency-groups]
dev = [
    "pytest",
    "httpx",
    "ruff",
    "black",
    "isort",
]
debug = [
    "debugpy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server_tavily"]


# --- Tool Configurations (Optional but recommended) ---

[tool.ruff]
line-length = 88
target-version = "py312"
select = ["E", "F", "I", "W", "C90", "N", "D"]
ignore = ["D100", "D104", "D107", "D203", "D212"] 

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 88